feat(content): add WebReg links under index in CSP
This adds clickable WebReg links under the indexes on the CSP
page which redirect to Webreg, automatically selecting the term
the course is in and inputing the index of the course. The user
then only has to submit. This improves the course reg process by
reducing manual input and time.